requirements.txt
aiohttp==3.9.5
aiohttp_cors==0.7.0
aiosqlite==0.20.0
discord.py==2.4.0
Flask==3.0.3
flask_cors==4.0.1
matplotlib==3.9.1
Pillow==10.4.0
python-dotenv==1.0.1
pytz==2024.1
Requests==2.32.3
